Refactor prezzo veicoli: Integer + format migliaia (UX)

+++ src/main/java/com/example/dto/AutomobileDTO.java
@@ -11,7 +11,7 @@ public class AutomobileDTO {
     private String marca;
     private String codiceImmatricolazione;
     private LocalDate dataImmatricolazione;
-    private Double prezzo;
+    private Integer prezzo;
 
     // campi specifici Automobile
     private int porte;
