+++ src/main/java/com/example/repository/FurgoneRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface FurgoneRepository extends JpaRepository<Furgone, Long> {
-    List<Furgone> findByPrezzoLessThan(Double prezzo);
+    List<Furgone> findByPrezzoLessThan(Integer prezzo);
     List<Furgone> findByMarca(String marca);
 }
